refactor: shared app layer + Match ROM prefix caching (#2)
* refactor: extract shared app layer (UART/clock/init) so the demos stay minimal

- new inc/app.h + src/app.c: lossless UART TX ring buffer, app_init()
  bundles clock+UART+LED setup, safe uart_write_int (full int32 range),
  ds18b20_busy strong override
- demo.c/demo2.c: only demo logic remains (callback + main), UART is now
  lossless in both, banners flushed before main loop
- Makefile: add src/app.c, per-app UART_TX_BUF_SIZE (demo 128, demo2 256)

* perf: cache the invariant Match ROM prefix in ds18b20_select

build_addr_cmd() re-encoded 0x55 + 8-byte ROM (72 slots) twice per
measurement cycle although it only depends on the selected device.
Now ds18b20_select() builds the prefix once (build_addr_prefix) and
build_addr_cmd() only encodes the varying command byte (8 slots).
